Split Amrap classes into different files.
